requirements.txt
fpdf==1.7.2
opencv_python==4.5.1.48
numpy==1.20.0
Pillow==8.2.0
pytesseract==0.3.8
